@@ -79,7 +79,7 @@ public class DefaultTbServiceInfoProvider implements TbServiceInfoProvider {
             serviceTypes = Collections.singletonList(ServiceType.of(serviceType));
         }
 
-        serviceInfo = getServiceInfoWithCurrentSystemInfo();
+       generateNewServiceInfoWithCurrentSystemInfo();
     }
 
     @AfterContextReady
@@ -106,13 +106,13 @@ public class DefaultTbServiceInfoProvider implements TbServiceInfoProvider {
     }
 
     @Override
-    public ServiceInfo getServiceInfoWithCurrentSystemInfo() {
+    public ServiceInfo generateNewServiceInfoWithCurrentSystemInfo() {
         ServiceInfo.Builder builder = ServiceInfo.newBuilder()
                 .setServiceId(serviceId)
                 .addAllServiceTypes(serviceTypes.stream().map(ServiceType::name).collect(Collectors.toList()))
                 .setSystemInfo(getCurrentSystemInfoProto());
 
-        return builder.build();
+        return serviceInfo = builder.build();
     }
 
     private TransportProtos.SystemInfoProto getCurrentSystemInfoProto() {
