@@ -16,7 +16,6 @@
 package org.thingsboard.server.queue.discovery;
 
 import org.thingsboard.server.common.msg.queue.ServiceType;
-import org.thingsboard.server.gen.transport.TransportProtos;
 import org.thingsboard.server.gen.transport.TransportProtos.ServiceInfo;
 
 public interface TbServiceInfoProvider {
@@ -29,6 +28,6 @@ public interface TbServiceInfoProvider {
 
     boolean isService(ServiceType serviceType);
 
-    ServiceInfo getServiceInfoWithCurrentSystemInfo();
+    ServiceInfo generateNewServiceInfoWithCurrentSystemInfo();
 
 }
