@@ -14,7 +14,7 @@
 /// limitations under the License.
 ///
 
-import { Component, forwardRef, Input, OnInit } from '@angular/core';
+import { Component, forwardRef, Input, OnDestroy, OnInit } from '@angular/core';
 import {
   AbstractControl,
   ControlValueAccessor,
@@ -28,7 +28,7 @@ import {
   Validator,
   Validators
 } from '@angular/forms';
-import { Observable, Subscription } from 'rxjs';
+import { Observable, Subject } from 'rxjs';
 import {
   EntityKeyType,
   entityKeyTypeTranslationMap,
@@ -39,6 +39,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { deepClone } from '@core/utils';
 import { KeyFilterDialogComponent, KeyFilterDialogData } from '@home/components/filter/key-filter-dialog.component';
 import { EntityId } from '@shared/models/id/entity-id';
+import { takeUntil } from 'rxjs/operators';
 
 @Component({
   selector: 'tb-key-filter-list',
@@ -57,7 +58,7 @@ import { EntityId } from '@shared/models/id/entity-id';
     }
   ]
 })
-export class KeyFilterListComponent implements ControlValueAccessor, Validator, OnInit {
+export class KeyFilterListComponent implements ControlValueAccessor, Validator, OnInit, OnDestroy {
 
   @Input() disabled: boolean;
 
@@ -75,22 +76,30 @@ export class KeyFilterListComponent implements ControlValueAccessor, Validator, 
 
   keyFiltersControl: UntypedFormControl;
 
+  private destroy$ = new Subject<void>();
   private propagateChange = null;
-
-  private valueChangeSubscription: Subscription = null;
 
   constructor(private fb: UntypedFormBuilder,
               private dialog: MatDialog) {
   }
 
   ngOnInit(): void {
-    this.keyFilterListFormGroup = this.fb.group({});
-    this.keyFilterListFormGroup.addControl('keyFilters',
-      this.fb.array([]));
+    this.keyFilterListFormGroup = this.fb.group({
+      keyFilters: this.fb.array([])
+    });
     this.keyFiltersControl = this.fb.control(null);
+
+    this.keyFilterListFormGroup.valueChanges.pipe(
+      takeUntil(this.destroy$)
+    ).subscribe(() => this.updateModel());
   }
 
-  keyFiltersFormArray(): UntypedFormArray {
+  ngOnDestroy() {
+    this.destroy$.next();
+    this.destroy$.complete();
+  }
+
+  get keyFiltersFormArray(): UntypedFormArray {
     return this.keyFilterListFormGroup.get('keyFilters') as UntypedFormArray;
   }
 
@@ -119,23 +128,21 @@ export class KeyFilterListComponent implements ControlValueAccessor, Validator, 
   }
 
   writeValue(keyFilters: Array<KeyFilterInfo>): void {
-    if (this.valueChangeSubscription) {
-      this.valueChangeSubscription.unsubscribe();
-    }
-    const keyFilterControls: Array<AbstractControl> = [];
-    if (keyFilters) {
-      for (const keyFilter of keyFilters) {
-        keyFilterControls.push(this.fb.control(keyFilter, [Validators.required]));
-      }
-    }
-    this.keyFilterListFormGroup.setControl('keyFilters', this.fb.array(keyFilterControls));
-    this.valueChangeSubscription = this.keyFilterListFormGroup.valueChanges.subscribe(() => {
-      this.updateModel();
-    });
-    if (this.disabled) {
-      this.keyFilterListFormGroup.disable({emitEvent: false});
+    if (keyFilters.length === this.keyFiltersFormArray.length) {
+      this.keyFiltersFormArray.patchValue(keyFilters, {emitEvent: false});
     } else {
-      this.keyFilterListFormGroup.enable({emitEvent: false});
+      const keyFilterControls: Array<AbstractControl> = [];
+      if (keyFilters) {
+        for (const keyFilter of keyFilters) {
+          keyFilterControls.push(this.fb.control(keyFilter, [Validators.required]));
+        }
+      }
+      this.keyFilterListFormGroup.setControl('keyFilters', this.fb.array(keyFilterControls), {emitEvent: false});
+      if (this.disabled) {
+        this.keyFilterListFormGroup.disable({emitEvent: false});
+      } else {
+        this.keyFilterListFormGroup.enable({emitEvent: false});
+      }
     }
     const keyFiltersArray = keyFilterInfosToKeyFilters(keyFilters);
     this.keyFiltersControl.patchValue(keyFiltersArray, {emitEvent: false});
