@@ -27,7 +27,7 @@ import {
   Validators
 } from '@angular/forms';
 import { SnmpCommunicationConfig, SnmpSpecType, SnmpSpecTypeTranslationMap } from '@shared/models/device.models';
-import { Subject, Subscription } from 'rxjs';
+import { Subject } from 'rxjs';
 import { isUndefinedOrNull } from '@core/utils';
 import { takeUntil } from 'rxjs/operators';
 
@@ -58,7 +58,6 @@ export class SnmpDeviceProfileCommunicationConfigComponent implements OnInit, On
   disabled: boolean;
 
   private usedSpecType: SnmpSpecType[] = [];
-  private valueChange$: Subscription = null;
   private destroy$ = new Subject<void>();
   private propagateChange = (v: any) => { };
 
@@ -68,17 +67,17 @@ export class SnmpDeviceProfileCommunicationConfigComponent implements OnInit, On
     this.deviceProfileCommunicationConfig = this.fb.group({
       communicationConfig: this.fb.array([])
     });
+    this.deviceProfileCommunicationConfig.valueChanges.pipe(
+      takeUntil(this.destroy$)
+    ).subscribe(() => this.updateModel());
   }
 
   ngOnDestroy() {
-    if (this.valueChange$) {
-      this.valueChange$.unsubscribe();
-    }
     this.destroy$.next();
     this.destroy$.complete();
   }
 
-  communicationConfigFormArray(): UntypedFormArray {
+  get communicationConfigFormArray(): UntypedFormArray {
     return this.deviceProfileCommunicationConfig.get('communicationConfig') as UntypedFormArray;
   }
 
@@ -99,27 +98,27 @@ export class SnmpDeviceProfileCommunicationConfigComponent implements OnInit, On
   }
 
   writeValue(communicationConfig: SnmpCommunicationConfig[]) {
-    if (this.valueChange$) {
-      this.valueChange$.unsubscribe();
-    }
-    const communicationConfigControl: Array<AbstractControl> = [];
-    if (communicationConfig) {
-      communicationConfig.forEach((config) => {
-        communicationConfigControl.push(this.createdFormGroup(config));
-      });
-    }
-    this.deviceProfileCommunicationConfig.setControl('communicationConfig', this.fb.array(communicationConfigControl));
-    if (!communicationConfig || !communicationConfig.length) {
-      this.addCommunicationConfig();
-    }
-    if (this.disabled) {
-      this.deviceProfileCommunicationConfig.disable({emitEvent: false});
+    if (communicationConfig?.length === this.communicationConfigFormArray.length) {
+      this.communicationConfigFormArray.patchValue(communicationConfig, {emitEvent: false});
     } else {
-      this.deviceProfileCommunicationConfig.enable({emitEvent: false});
+      const communicationConfigControl: Array<AbstractControl> = [];
+      if (communicationConfig) {
+        communicationConfig.forEach((config) => {
+          communicationConfigControl.push(this.createdFormGroup(config));
+        });
+      }
+      this.deviceProfileCommunicationConfig.setControl(
+        'communicationConfig', this.fb.array(communicationConfigControl), {emitEvent: false}
+      );
+      if (!communicationConfig || !communicationConfig.length) {
+        this.addCommunicationConfig();
+      }
+      if (this.disabled) {
+        this.deviceProfileCommunicationConfig.disable({emitEvent: false});
+      } else {
+        this.deviceProfileCommunicationConfig.enable({emitEvent: false});
+      }
     }
-    this.valueChange$ = this.deviceProfileCommunicationConfig.valueChanges.subscribe(() => {
-      this.updateModel();
-    });
     this.updateUsedSpecType();
     if (!this.disabled && !this.deviceProfileCommunicationConfig.valid) {
       this.updateModel();
@@ -133,16 +132,16 @@ export class SnmpDeviceProfileCommunicationConfigComponent implements OnInit, On
   }
 
   public removeCommunicationConfig(index: number) {
-    this.communicationConfigFormArray().removeAt(index);
+    this.communicationConfigFormArray.removeAt(index);
   }
 
 
   get isAddEnabled(): boolean {
-    return this.communicationConfigFormArray().length !== Object.keys(SnmpSpecType).length;
+    return this.communicationConfigFormArray.length !== Object.keys(SnmpSpecType).length;
   }
 
   public addCommunicationConfig() {
-    this.communicationConfigFormArray().push(this.createdFormGroup());
+    this.communicationConfigFormArray.push(this.createdFormGroup());
     this.deviceProfileCommunicationConfig.updateValueAndValidity();
     if (!this.deviceProfileCommunicationConfig.valid) {
       this.updateModel();
